Basic data fetch from db to display user stats
