+++ app/Models/User.php
@@ -7,6 +7,7 @@ namespace App\Models;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
@@ -41,6 +42,11 @@ class User extends Authenticatable
         return $this->belongsToMany(Task::class, "user_tasks")
             ->withPivot("is_completed", "completed_at", "progress")
             ->withTimestamps();
+    }
+
+    public function userStatistics(): HasOne
+    {
+        return $this->hasOne(UserStatistics::class);
     }
 
     protected function casts(): array
